+++ client/src/api/index.ts
@@ -1,8 +1,10 @@
-// Poll API exports
+// Queries exports
 export * from './queries/poll';
 export * from './queries/thread';
+export * from './queries/notification';
 
 // Hook exports
 export * from './hooks/usePoll';
 export * from './hooks/useThreads';
-export * from './hooks/useCreatePost'; 
+export * from './hooks/useCreatePost';
+export * from './hooks/useNotifications';
